rust: io: refactor io to use IoAccess trait
adds the traits IoAccess and IoAccess64 and refactors Io
to use them. also make Io use the ioread/iowrite functions
instead of the read*/write* functions.

this fixes a problem with `pci::Bar` which used Io to access
an address from pci_iomap(). addresses from pci_iomap() can be
either MMIO or PIO so it. should be accessed with the ioread/iowrite
functions.

Link: https://www.kernel.org/doc/html/v6.15-rc1/driver-api/pci/pci.html#c.pci_iomap
Signed-off-by: Fiona Behrens <me@kloenk.dev>
Co-developed-by: Andrew Ballance <andrewjballance@gmail.com>
Signed-off-by: Andrew Ballance <andrewjballance@gmail.com>