WIP: add enum IO type
adds new IO enum that is either a MMIo or PortIo.

adds from_raw_cookie methods which builds an io from
an address that should be accessed by ioread/iowrite

refactor pci::Bar to use the new Io type. pci::bar
has a new field, original_ioptr because the decoded
iomem address might not be the same as the original address
and bad needs that for unmapping the memory.

will split this into multiple commits before submission